fixed unauthorized problems in recipes and users

+++ src/recipes/recipes.controller.ts
@@ -10,6 +10,8 @@ import {
   Patch,
   Post,
   Query,
+  Req,
+  UnauthorizedException,
   UseGuards,
   ValidationPipe,
 } from '@nestjs/common';
@@ -19,6 +21,7 @@ import { CreateRecipeDto } from './dto/createRecipe.dto';
 import { UpdateRecipeDto } from './dto/updateRecipe.dto';
 import { Prisma } from 'generated/prisma';
 import { JwtGuard } from 'src/auth/guards/jwt.guard';
+import { Request } from 'express';
 
 @Controller('recipes')
 @UseGuards(JwtGuard)
@@ -43,7 +46,16 @@ export class RecipesController {
   }
 
   @Post()
-  async create(@Body(ValidationPipe) createRecipeDto: CreateRecipeDto) {
+  async create(
+    @Body(ValidationPipe) createRecipeDto: CreateRecipeDto,
+    @Req() req: Request,
+  ) {
+    if ((req.user as { id: number }).id !== createRecipeDto.authorId) {
+      throw new UnauthorizedException(
+        'Only user can post recepies with his authorID',
+      );
+    }
+
     const newRecipe = await this.recipesService.create(createRecipeDto);
 
     if (!newRecipe) throw new BadRequestException();
@@ -55,8 +67,13 @@ export class RecipesController {
   async update(
     @Param('id', ParseIntPipe) id: number,
     @Body(ValidationPipe) updateRecipeDto: UpdateRecipeDto,
+    @Req() req: Request,
   ) {
     try {
+      if ((req.user as { id: number }).id !== updateRecipeDto.authorId) {
+        throw new UnauthorizedException('Only user can chacnge their recipes');
+      }
+
       await this.getOne(id);
 
       if (!Object.keys(updateRecipeDto).length) {
@@ -78,8 +95,12 @@ export class RecipesController {
   }
 
   @Delete(':id')
-  async delete(@Param('id', ParseIntPipe) id: number) {
-    await this.getOne(id);
+  async delete(@Param('id', ParseIntPipe) id: number, @Req() req: Request) {
+    const recipe = await this.getOne(id);
+
+    if ((req.user as { id: number }).id !== recipe.authorId) {
+      throw new UnauthorizedException('Only user can delete their recipes');
+    }
 
     const result = await this.recipesService.remove(id);
